+++ src/sstc_core/sites/spectral/data_products/qflags.py
@@ -4,8 +4,7 @@ from sstc_core.sites.spectral import utils
 def compute_qflag(
     latitude_dd: float,
     longitude_dd: float, 
-    records_dict: dict,
-    has_snow_presence: bool = False,
+    records_dict: dict,    
     timezone_str: str = 'Europe/Stockholm',
     is_per_image: bool = False, 
     default_temporal_resolution: bool = True,  # default temporal resolution is 30 min, else 1 hr or more
@@ -15,7 +14,7 @@ def compute_qflag(
     `is_per_image` if True the len of records_dict will not be considered on the qflag.
     """
     
-    datetime_list = [v['creation_date'] for k, v in records_dict.items()]
+    datetime_list = [v['creation_date'] for _, v in records_dict.items()]
     
     mean_datetime_str = utils.mean_datetime_str(datetime_list=datetime_list)
     sun_position = utils.calculate_sun_position(
@@ -29,8 +28,7 @@ def compute_qflag(
     solar_elevation_class = utils.get_solar_elevation_class(sun_elevation=sun_elevation_angle)
    
     n_records = len(records_dict)
-    
-    
+        
     if (n_records < 3 if default_temporal_resolution else 2) and (solar_elevation_class == 1):
         QFLAG = 11
         if not is_per_image:
@@ -44,16 +42,14 @@ def compute_qflag(
             weight: 0.5
         else:
             weight: 0.75
-        
-    
+            
     elif (n_records < 3 if default_temporal_resolution else 2) and (solar_elevation_class == 3):
         QFLAG = 13
         if not is_per_image:
             weight: 0.5
         else:
             weight: 1
-        
-    
+            
     elif ((n_records >= 3 if default_temporal_resolution else 2) and (n_records < 6 if default_temporal_resolution else 4 )) and (solar_elevation_class == 1):
         QFLAG = 21
         weight: 0.5
@@ -63,8 +59,7 @@ def compute_qflag(
         QFLAG = 22
         weight: 0.75
   
-        
-        
+                
     elif ((n_records >= 3 if default_temporal_resolution else 2) and (n_records < 6 if default_temporal_resolution else 4)) and (solar_elevation_class == 3):
         QFLAG = 23
         weight: 1